+++ cr_channable_connector/doc/changelog.rst
@@ -2,3 +2,11 @@
 -------------------------------
 
 - migrated module from version 19.0
+
+18.0.0.1(Date: 10th Aug,2026)
+-------------------------------
+
+- Update existing Odoo orders seamlessly during bulk sync processes.
+- Automatic Odoo order cancellation and credit note creation when a bulk sync detects a cancelled status in Channable.
+- Added 'Schedule Sync' smart button to the Channable Marketplace dashboard for easy sync interval configuration.
+- Removed 'Sync State' smart button because similar functionality like 'Sync Order'.